+++ test/http/GraphQLHttpServer.test.ts
@@ -11,16 +11,13 @@ describe("The GraphQLHttpServer", () => {
   });
 
   it("should be able to handle POST GraphQL requests", async () => {
-    const response = await gqlServer.http.execute(
-      {
-        method: "POST",
-        headers: {},
-        body: {
-          query: "{ praise }",
-        },
+    const response = await gqlServer.http.execute({
+      method: "POST",
+      headers: {},
+      body: {
+        query: "{ praise }",
       },
-      undefined
-    );
+    });
 
     expect(response).toEqual({
       status: 200,
@@ -34,28 +31,22 @@ describe("The GraphQLHttpServer", () => {
   });
 
   it("should be able to handle GET GraphQL requests", async () => {
-    const response = await gqlServer.http.execute(
-      {
-        method: "GET",
-        headers: {},
-        query: {
-          query: `{ praise }`,
-        },
+    const response = await gqlServer.http.execute({
+      method: "GET",
+      headers: {},
+      query: {
+        query: `{ praise }`,
       },
-      undefined
-    );
+    });
 
     // test query caching
-    await gqlServer.http.execute(
-      {
-        method: "GET",
-        headers: {},
-        query: {
-          query: `{ praise }`,
-        },
+    await gqlServer.http.execute({
+      method: "GET",
+      headers: {},
+      query: {
+        query: `{ praise }`,
       },
-      undefined
-    );
+    });
 
     expect(response).toEqual({
       status: 200,
@@ -69,17 +60,14 @@ describe("The GraphQLHttpServer", () => {
   });
 
   it("should reject unsupported methods", async () => {
-    const response = await gqlServer.http.execute(
-      {
-        method: "PUT",
-        headers: {},
-        body: {
-          query: `{ praise }`,
-          operationName: "String",
-        },
+    const response = await gqlServer.http.execute({
+      method: "PUT",
+      headers: {},
+      body: {
+        query: `{ praise }`,
+        operationName: "String",
       },
-      undefined
-    );
+    });
 
     expect(response).toEqual({
       status: 405,
@@ -95,17 +83,14 @@ describe("The GraphQLHttpServer", () => {
   });
 
   it("should reject mutations via GET", async () => {
-    const response = await gqlServer.http.execute(
-      {
-        method: "GET",
-        headers: {},
-        query: {
-          query: "mutation Test { youDied }",
-          operationName: "Test",
-        },
+    const response = await gqlServer.http.execute({
+      method: "GET",
+      headers: {},
+      query: {
+        query: "mutation Test { youDied }",
+        operationName: "Test",
       },
-      undefined
-    );
+    });
 
     expect(response).toEqual({
       status: 400,
@@ -121,16 +106,13 @@ describe("The GraphQLHttpServer", () => {
   });
 
   it("should reject invalid queries", async () => {
-    const response = await gqlServer.http.execute(
-      {
-        method: "POST",
-        headers: {},
-        body: {
-          query: `{ invalid }`,
-        },
+    const response = await gqlServer.http.execute({
+      method: "POST",
+      headers: {},
+      body: {
+        query: `{ invalid }`,
       },
-      undefined
-    );
+    });
 
     expect(response).toEqual({
       status: 400,